move the google places to a different folder
