+++ lib/delivery_map/google_places.ex
@@ -1,0 +1,50 @@
+defmodule DeliveryMap.GooglePlaces do
+  @moduledoc """
+  Helper module for Google Places API integration.
+  """
+  @autocomplete_url "https://maps.googleapis.com/maps/api/place/autocomplete/json"
+  @details_url "https://maps.googleapis.com/maps/api/place/details/json"
+
+  def autocomplete(query) do
+    key = Application.get_env(:delivery_map, :google_maps_api_key) || "YOUR_API_KEY"
+
+    params =
+      URI.encode_query(%{
+        input: query,
+        key: key
+      })
+
+    url = "#{@autocomplete_url}?#{params}"
+
+    case Req.get(url) do
+      {:ok, %{body: %{"predictions" => preds}}} ->
+        Enum.map(preds, fn pred ->
+          %{description: pred["description"], place_id: pred["place_id"]}
+        end)
+
+      _something_else ->
+        []
+    end
+  end
+
+  def get_address(place_id) do
+    key = Application.get_env(:delivery_map, :google_maps_api_key) || "YOUR_API_KEY"
+
+    params =
+      URI.encode_query(%{
+        place_id: place_id,
+        key: key,
+        fields: "formatted_address"
+      })
+
+    url = "#{@details_url}?#{params}"
+
+    case Req.get(url) do
+      {:ok, %{body: %{"result" => %{"formatted_address" => addr}}}} ->
+        addr
+
+      _ ->
+        nil
+    end
+  end
+end
